reports/python/requirements.txt
pandas==2.0.3
openpyxl==3.1.2
psycopg2-binary==2.9.7
